made Get request using Retrofit, lesson 47

+++ app/src/main/java/kg/tutorialapp/wheather_final_project/ForeCast.kt
@@ -36,7 +36,7 @@ data class HourlyForeCast(
     var temp: Double? = null,
     var weather: List<Weather>? = null,
     @SerializedName("pop")
-    var probability: Int? = null
+    var probability: Double? = null
 )
 
 data class DailyForeCast(
@@ -45,7 +45,7 @@ data class DailyForeCast(
     var temp: Temperature? = null,
     var weather: List<Weather>? = null,
     @SerializedName("pop")
-    var probability: Int? = null
+    var probability: Double? = null
 
 )
